indexer/requirements.txt
asyncpg
meilisearch
python-dotenv
minio
beautifulsoup4
lxml